Atualização do Sistema de Dia e Noite
Basicamente eu fiz a interface voltar a funcionar, as horas e os dias passam e as fases da lua também são trocadas.
