+++ Assets/Materials/SkyMaterial.mat
@@ -25,11 +25,11 @@ Material:
     m_Ints: []
     m_Floats:
     - _AtmosphereThickness: 0.7
-    - _Exposure: 0.10287131
+    - _Exposure: 1.159098
     - _SunDisk: 0
     - _SunSize: 0.04
     - _SunSizeConvergence: 5
     m_Colors:
-    - _GroundColor: {r: 0.0060074753, g: 0.027025167, b: 0.19215687, a: 0.5}
-    - _SkyTint: {r: 0.012014951, g: 0.054050334, b: 0.38431373, a: 1}
+    - _GroundColor: {r: 0.45754716, g: 0.4480255, b: 0.40359113, a: 0.5}
+    - _SkyTint: {r: 0.9150943, g: 0.896051, b: 0.80718225, a: 1}
   m_BuildTextureStacks: []
