--- Assets/Scripts/Controllers/TimeController_Old.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 46ff3fb8540a8b54d8b34438bfe42aba
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
